requirements.txt
# Core dependencies
catanatron>=3.2.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Web visualization
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Data processing and storage
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Elo rating system
trueskill>=0.4.5

# Logging and monitoring
loguru>=0.7.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0
pre-commit>=3.5.0